requirements.txt

google-adk>=0.1.0
google-generativeai>=0.8.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
chromadb>=0.5.0
python-dotenv>=1.0.0
streamlit>=1.40.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
requests>=2.32.0
